+++ app/Widgets/AvatarPlaceholder/AvatarPlaceholder.php
@@ -6,8 +6,7 @@ class AvatarPlaceholder extends \Movim\Widget\Base
 {
     public function display()
     {
-        $id = urldecode($this->get('id'));
-        $size = $this->get('size');
+        $id = urldecode($this->get('id') ?? 'M');
 
         $letters = firstLetterCapitalize(slugify($id));
         header_remove('Content-Type');