+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Movim Changelog
 
 v0.30 (master)
 ---------------------------
+* Limit the jid, name and group name to 256 character max in the Roster to allow it to be saved in the DB
 
 v0.29.2
 ---------------------------